moved lir/data/data_strategies.py to lir/data_strategies.py
