+++ lir/data_strategies.py
