+++ tests/lrsystems/test_specific_source.py
@@ -4,8 +4,8 @@ import numpy as np
 import pytest
 from sklearn.preprocessing import StandardScaler
 
-from lir.data.data_strategies import BinaryTrainTestSplit
 from lir.data.models import FeatureData
+from lir.data_strategies import BinaryTrainTestSplit
 from lir.lrsystems.binary_lrsystem import BinaryLRSystem
 from lir.lrsystems.lrsystems import LLRData
 from lir.transform.pipeline import Pipeline
