+++ tests/test_data_strategies.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-from lir.data.data_strategies import MulticlassTrainTestSplit, PairedInstancesTrainTestSplit, PredefinedTrainTestSplit
 from lir.data.datasets.synthesized_normal_multiclass import (
     SynthesizedDimension,
     SynthesizedNormalMulticlassData,
 )
 from lir.data.models import FeatureData
+from lir.data_strategies import MulticlassTrainTestSplit, PairedInstancesTrainTestSplit, PredefinedTrainTestSplit
 from lir.transform.pairing import InstancePairing
